Some fixes, remove redundant as_dict code and constructors
