+++ backend/api/model/api.py
@@ -4,12 +4,12 @@ One to many relationship with calls. Collection.
 '''
 from database import db
 import datetime
-from base import Base
+from .base import Base
 
 class API(Base):
     __tablename__ = 'apis'
 
-    callees = db.relationship('calls', backref='api', lazy='dynamic')
+    callees = db.relationship('Call', backref='API', lazy='dynamic')
 
     def __repr__(self):
         return '<API %r>' % self.name
