+++ backend/api/model/call.py
@@ -4,11 +4,15 @@ API/language, and be related to many posts containing usage of this call. This t
 searched against.
 '''
 from .base import Base
-from ..app import app, db
+from database import db
 
 class Call(Base):
     __tablename__ = 'calls'
     id = db.Column(db.Integer, primary_key=True)
     # name should be unique
     full_name = db.Column(db.String(100), nullable=False)
-    api = db.relationship('API', backref=db.backref('posts', lazy=True))
+    api_id = db.Column(db.Integer, db.ForeignKey('apis.id'), nullable=False)
+
+
+    def __repr__(self):
+        return '<Call %r>' % self.full_name
