+++ backend/app.py
@@ -24,6 +24,17 @@ def create_app():
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
     db.init_app(app)
+
+    # add test models to db
+    from api.model.api import API
+    from api.model.post import Post
+    from api.model.call import Call
+    
+    api = API(name='test_api')
+    post = Post(content='test_post')
+    call = Call(full_name='test_call')
+
+    db.create_all()
     
     @app.before_first_request
     def create_table():
